refactor: extract metadata extraction prompt value
